// backend/app/src/gen/pb/api/v1/app/room/play/service/PlayService_rb.service.ts
// @generated by protoc-gen-restobuf-nestjs v1 with parameter "target=ts,import_extension=js"
// @generated from file api/v1/app/room/play/service.proto (package api.v1.app.room, syntax proto3)
/* eslint-disable */

import { Injectable } from "@nestjs/common";
import type { AddDeckRequest, AddDeckResponse, CreatePlaceRequest, CreatePlaceResponse, DeletePlaceRequest, DeletePlaceResponse, GetHistoryRequest, GetHistoryResponse, ListPlaceRequest, ListPlaceResponse, MoveCardRequest, MoveCardResponse } from "../service_pb.js";
import { Request, Response } from "express";

@Injectable() 
export abstract class PlayServiceService {
  abstract handleListPlace(input: ListPlaceRequest, req: Request, res: Response): Promise<ListPlaceResponse>;
  abstract handleCreatePlace(input: CreatePlaceRequest, req: Request, res: Response): Promise<CreatePlaceResponse>;
  abstract handleDeletePlace(input: DeletePlaceRequest, req: Request, res: Response): Promise<DeletePlaceResponse>;
  abstract handleAddDeck(input: AddDeckRequest, req: Request, res: Response): Promise<AddDeckResponse>;
  abstract handleMoveCard(input: MoveCardRequest, req: Request, res: Response): Promise<MoveCardResponse>;
  abstract handleGetHistory(input: GetHistoryRequest, req: Request, res: Response): Promise<GetHistoryResponse>;
}